ML/ml-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
py-eureka-client==0.11.11
PyJWT==2.8.0
python-keycloak==3.3.0
httpx==0.25.2
pytest==7.4.3
python-multipart==0.0.6
python-dotenv==1.0.0
# Machine Learning stubs
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3
